frontend/requirements.txt
streamlit>=1.28.0
requests==2.31.0
rdkit-pypi==2023.9.1
pubchempy==1.0.4
pandas==2.1.3
numpy==1.24.3
matplotlib==3.7.2
streamlit-aggrid==0.3.4
py3Dmol==2.0.4
stmol==0.0.6 